Refactor request error
