+++ functions/src/controllers/errors/RequestError.ts
@@ -1,0 +1,14 @@
+type ErrorCode =
+  | 'user-needs-email'
+  | 'request-not-found'
+  | 'request-exists'
+  | 'request-expired'
+  | 'verification-failed';
+export class RequestError extends Error {
+  constructor(code: ErrorCode) {
+    super(code);
+    this.code = code;
+  }
+
+  code: string;
+}
