--- functions/src/controllers/errors/requestsError.ts
@@ -1,15 +1,0 @@
-export class RequestError extends Error {
-  constructor(code: ErrorCode) {
-    super(code);
-    this.code = code;
-  }
-
-  code: string;
-}
-
-export type ErrorCode =
-  | 'user-needs-email'
-  | 'request-not-found'
-  | 'request-exists'
-  | 'request-expired'
-  | 'verification-failed';
