+++ functions/src/models/types/types.ts
@@ -3,4 +3,6 @@ import {Timestamp} from 'firebase-admin/firestore';
 export type RequestData = {
   verificationCode: number;
   expires: Timestamp;
+  createdAt: Timestamp;
+  updatedAt: Timestamp;
 };
